week6/interview-analysis/prop_tax_assesment_ext/backend/input_validator/pyproject.toml
[project]
name = "input-validator"
version = "0.1.0"
description = "Property tax document input file validation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "strands-agents=1.15.0",
]
